Raspberry Pi Files
Code that is currently on the Pi for the SSTV stuff.  The sstvcam.sh
file is the one that currently transmits the images taken, the other .sh
in the pisstv folder is primarily for testing.  The pifm.c code is in
the pisstv folder as well.  I changed the frequency and some clock
properties in that.